# [snapshot] Avoid string allocations