--- a/pkg/backend/httpstate/diffs.go
+++ b/pkg/backend/httpstate/diffs.go
@@ -22,6 +22,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"sync"
+	"unsafe"
 
 	"github.com/hexops/gotextdiff/myers"
 	"github.com/hexops/gotextdiff/span"
@@ -30,6 +31,15 @@ import (
 
 	opentracing "github.com/opentracing/opentracing-go"
 )
+
+func asString(raw json.RawMessage) string {
+	// NOTE: this depends on the fact that the layout of a byte slice and a string share a common header. Both of these
+	// types are laid out as (ptr, len). Once the CLI targets Go 1.20, we should change this to
+	//
+	//     return unsafe.String(unsafe.SliceData([]byte(raw)), len(raw))
+	//
+	return *(*string)(unsafe.Pointer(&raw))
+}
 
 type deployment struct {
 	raw json.RawMessage
@@ -113,7 +123,7 @@ func (dds *deploymentDiffState) Diff(ctx context.Context, deployment deployment)
 		checkpointHash = dds.computeHash(childCtx, after)
 	}()
 
-	delta, err := dds.computeEdits(childCtx, string(before), string(after))
+	delta, err := dds.computeEdits(childCtx, asString(before), asString(after))
 	if err != nil {
 		return deploymentDiff{}, fmt.Errorf("Cannot marshal the edits: %v", err)
 	}