--- a/pkg/backend/httpstate/diffs.go
+++ b/pkg/backend/httpstate/diffs.go
@@ -32,7 +32,11 @@ import (
 	opentracing "github.com/opentracing/opentracing-go"
 )
 
-func asString(raw json.RawMessage) string {
+// unsafeAsString converts the contents of the given json.RawMessage into a string without allocating. The contents
+// of the returned string will change if the json.RawMessage is mutated before the referenced string goes out of scope.
+// It is imperative that callers ensure that the lifetime of the return value does not exceed the lifetime of the
+// input contents.
+func unsafeAsString(raw json.RawMessage) string {
 	// NOTE: this depends on the fact that the layout of a byte slice and a string share a common header. Both of these
 	// types are laid out as (ptr, len). Once the CLI targets Go 1.20, we should change this to
 	//
@@ -123,7 +127,7 @@ func (dds *deploymentDiffState) Diff(ctx context.Context, deployment deployment)
 		checkpointHash = dds.computeHash(childCtx, after)
 	}()
 
-	delta, err := dds.computeEdits(childCtx, asString(before), asString(after))
+	delta, err := dds.computeEdits(childCtx, unsafeAsString(before), unsafeAsString(after))
 	if err != nil {
 		return deploymentDiff{}, fmt.Errorf("Cannot marshal the edits: %v", err)
 	}