@pgavlin 2023-05-16
abandoning in favor of https://github.com/pulumi/pulumi/pull/12962